i45: Combine map hooks and clean up WorldMap
