+++ src/library/hooks/index.js
@@ -1,3 +1,2 @@
 export { default as useLocation } from './useLocation';
-export { default as useMapCamera } from './useMapCamera';
-export { default as useMapStorage } from './useMapStorage';
+export { default as useMap } from './useMap';
